feat(appcomposer): load and save from VS Code buffer
## Problem
- Application Composer works directly off of the workspace files.
Changes in VS Code will only be detected by App Composer when the user
saves, and changes in Composer will immediately be saved to disk.
This does not follow the standard VS Code pattern.

## Solution
- This updates Application Composer to read from and write to VS Code
text documents when possible, ensuring that Composer is synced with
the latest user changes and giving the user more control over what
changes are saved.